Move DocMapping to utility folder.
