+++ processor/error/event.go
@@ -51,8 +51,8 @@ func (e *Event) DocType() string {
 	return "error"
 }
 
-func (e *Event) Mappings(pa *payload) (time.Time, []m.DocMapping) {
-	mapping := []m.DocMapping{
+func (e *Event) Mappings(pa *payload) (time.Time, []utility.DocMapping) {
+	mapping := []utility.DocMapping{
 		{Key: "processor", Apply: func() common.MapStr {
 			return common.MapStr{"name": processorName, "event": e.DocType()}
 		}},
@@ -64,7 +64,7 @@ func (e *Event) Mappings(pa *payload) (time.Time, []m.DocMapping) {
 	}
 
 	if e.Transaction != nil {
-		mapping = append(mapping, m.DocMapping{
+		mapping = append(mapping, utility.DocMapping{
 			Key:   "transaction",
 			Apply: func() common.MapStr { return common.MapStr{"id": e.Transaction.Id} },
 		})
